+++ packages/authhero/test/routes/auth-api/token.spec.ts
@@ -500,7 +500,7 @@ describe("token", () => {
 
         const cookie = response.headers.get("set-cookie");
         expect(cookie).toBe(
-          `tenantId-auth-token=${accessToken?.payload.sid}; HttpOnly; Max-Age=604800; Path=/; SameSite=None; Secure`,
+          `tenantId-auth-token=${accessToken?.payload.sid}; HttpOnly; Max-Age=2592000; Path=/; SameSite=None; Secure`,
         );
       });
     });