+++ packages/adapter-interfaces/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @authhero/adapter-interfaces
 
+## 0.32.0
+
+### Minor Changes
+
+- add hooks to add claims to token
+
+## 0.31.0
+
+### Minor Changes
+
+- set used_at for codes
+
 ## 0.30.0
 
 ### Minor Changes